Change LIVE_LISTEN_URL_PATH to include /transcription at the base path (#1)
This is due to Contio's path-based routing in the load balancer.

The load balancer will see /transcription or /transcription/* as a listener rule and forward it to an EC2 instance.

That EC2 instance has Caddy that strips that off and then forwards it to the actual Deepgram API Docker container.